pyproject.toml
[tool.poetry]
name = "tap-pulumi-cloud"
version = "0.0.1"
description = "`tap-pulumi-cloud` is a Singer tap for Pulumi Cloud, built with the Meltano SDK for Singer Taps."
authors = ["Edgar Ramírez-Mondragón <edgarrm358@gmail.com>"]
keywords = [
  "ELT",
  "singer.io",
  "Pulumi Cloud",
]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/edgarrmondragon/tap-pulumi-cloud"
repository = "https://github.com/edgarrmondragon/tap-pulumi-cloud"
documentation = "https://github.com/edgarrmondragon/tap-pulumi-cloud#readme"

[tool.poetry.dependencies]
pyhumps = "^3.8.0"
python = "<3.12,>=3.7.1"
requests-cache = "^1.0.1"
singer-sdk = "0.24.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
use_parentheses = true
include_trailing_comma = true
src_paths = "tap_pulumi_cloud"
known_first_party = "tap_pulumi_cloud"
add_imports = [
  "from __future__ import annotations",
]

[tool.pytest.ini_options]
addopts = "-vvv"

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = ["backoff.*"]

[build-system]
requires = ["poetry-core>=1,<2"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
"tap-pulumi-cloud" = "tap_pulumi_cloud.tap:TapPulumiCloud.cli"

[tool.ruff]
ignore = [
  "ANN101", # missing-type-self
  "DJ",     # flake8-django
]
line-length = 88
select = ["ALL"]
src = ["tap_pulumi_cloud", "tests"]
target-version = "py37"

[tool.ruff.per-file-ignores]
"noxfile.py" = ["ANN"]
"tests/*" = [
  "ANN201",  # missing-return-type-public-function
  "S101",    # assert
  "SLF001",  # private-member-access
]

[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.isort]
known-first-party = ["tap_pulumi_cloud"]

[tool.ruff.pydocstyle]
convention = "google"
